pyproject.toml
[project]
name = "ai-engine"
version = "0.1.0"
description = "Behavioural intelligence engine for converged IoT/AI security"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "numpy>=1.26",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]
ml = [
    # filled in at step 4 (Isolation Forest)
    # "scikit-learn>=1.4",
    # "pandas>=2.2",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["schemas*", "dataset*"]

[tool.pytest.ini_options]
# `pythonpath = ["."]` puts the project root on sys.path so `from schemas import ...`
# works without any conftest.py hackery, and without installing the package.
pythonpath = ["."]
testpaths = ["tests", "dataset/tests"]
# Verbose output and short tracebacks are friendlier during development.
addopts = "-v --tb=short"
